training/requirements.txt
# CropNet training dependencies (GPU instance)
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.38.0
Pillow>=10.0.0
numpy>=1.26.0
scipy>=1.12.0
pyyaml>=6.0
tqdm>=4.66.0
onnx>=1.16.0
onnxruntime>=1.17.0
